refactor: correção no nome da pasta
