+++ alura-tracker/src/interfaces/ITarefa.ts
